requirements.txt
torch==2.9.1
transformers==4.57.3
opencv-python==4.12.0.88
pillow==12.0.0
